% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Launch interactive weather analysis app}
\usage{
app(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\description{
Launch interactive analysis of weather period comparison
for different RDWD stations.
The R session is blocked during usage, close the app to re-enable
console usage.
}
\examples{
# app()
}
\seealso{
\code{\link[shiny:runApp]{shiny::runApp()}}, \link{rdwd}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, July 2018 + April 2023
}
\keyword{iplot}
