% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checkSuggestedPackage}
\alias{checkSuggestedPackage}
\title{check suggested package for availability}
\usage{
checkSuggestedPackage(package, functionname)
}
\arguments{
\item{package}{Charstring: package to be checked for loadability}

\item{functionname}{Charstring: function name to be used in the message}
}
\value{
invisible success logical value from \code{\link[=requireNamespace]{requireNamespace()}}
}
\description{
check suggested package for availability,
yielding an instructive error message if not
}
\seealso{
\code{\link[=requireNamespace]{requireNamespace()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2019
}
\keyword{package}
