% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUpdates.R
\name{checkUpdates}
\alias{checkUpdates}
\title{check some historical files for updates by DWD}
\usage{
checkUpdates(fast = TRUE)
}
\arguments{
\item{fast}{only check a subset? currently ignored. DEFAULT: TRUE}
}
\value{
currently available files on the FTP server in the ca 34 historical folders, invisibly
}
\description{
check whether the DWD has updated historical datasets.
That requires \code{\link{updateIndexes}} to be run by me.
If that is the case, the funtion will give a warning, otherwise a message.
}
\examples{
current <- checkUpdates()
}
\seealso{
\url{https://bookdown.org/brry/rdwd/fileindex.html}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Mar 2025
}
