% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIndex.R
\name{createIndex}
\alias{createIndex}
\title{Create file and meta index of the DWD CDC FTP Server}
\usage{
createIndex(paths,
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  dir = "DWDdata", fname = "fileIndex.txt", meta = FALSE,
  metadir = "meta", mname = "metaIndex.txt", gname = "geoIndex.txt",
  quiet = FALSE, ...)
}
\arguments{
\item{paths}{Char: vector of DWD paths returned by \code{\link{indexDWD}} called
with the same \code{base} value as this function}

\item{base}{Main directory of DWD ftp server, defaulting to observed climatic records.
DEFAULT: \url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}}

\item{dir}{Char: writeable directory name where to save the main output(s).
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{fname}{Char: Name of file in \code{dir} in which to write \code{\link{fileIndex}}.
Use \code{fname=""} to suppress writing. DEFAULT: "fileIndex.txt"}

\item{meta}{Logical: should metaIndex also be created from fileIndex?
Uses \code{\link{dataDWD}} to download files if not present.
DEFAULT: FALSE}

\item{metadir}{Char: Directory (subfolder of \code{dir}) where original
description files are downloaded to if meta=TRUE. vPassed to
\code{\link{dataDWD}}. "" to write in \code{dir}. DEFAULT: "meta"}

\item{mname}{Char: Name of file in \code{dir} (not \code{metadir}) in which to
write \code{\link{metaIndex}}.
Use \code{mname=""} to suppress writing. DEFAULT: "metaIndex.txt"}

\item{gname}{Ditto for \code{\link{geoIndex}}. DEFAULT: "geoIndex.txt"}

\item{quiet}{Logical: Suppress messages about directory / filename? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{dataDWD}} for the meta part.}
}
\value{
invisible data.frame (or if meta=TRUE, list with two data.frames)
with a number of columns inferred from the paths. Each is also written to disc.
}
\description{
Create data.frames out of the vector index returned by \code{\link{indexDWD}}.
For \code{\link{fileIndex}} (the first output element) \code{createIndex}
tries to obtain res, var, pe,r file, id, start and end from the paths.
If \code{meta=TRUE}, \code{\link{metaIndex}} is also created, which combines
all Beschreibung files into a single data.frame.\cr
If you create your own index as suggested in selectDWD (argument \code{findex}),
you can read the produced file as shown in the example section.
}
\examples{
# see last part of
if(interactive())
browseURL("https://github.com/brry/rdwd/blob/master/R/rdwd-package.R")
# where fileIndex and metaIndex are added to the package

# Read results in later:
\dontrun{ ## files normally not yet available:
fileIndex2 <- read.table("DWDdata/fileIndex.txt", sep="\\t", header=TRUE,
                         colClasses="character")
metaIndex2 <- read.table("DWDdata/metaIndex.txt", sep="\\t", header=TRUE, as.is=TRUE)
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct-Nov 2016
}
\seealso{
\code{\link{indexDWD}}, \code{\link{fileIndex}}, \code{\link{metaIndex}}, \code{\link{selectDWD}}
}
\keyword{manip}

