% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDWD.R
\name{dataDWD}
\alias{dataDWD}
\title{Download data from the DWD CDC FTP Server}
\usage{
dataDWD(file, dir = "DWDdata", quiet = FALSE, read = TRUE, format = NA,
  sleep = 0, progbar = TRUE)
}
\arguments{
\item{file}{Complete file URL (including filename.zip) as returned by
\code{\link{metaDWD}} if \code{id} or \code{name} is given.
Can be a vector with several filenames.}

\item{dir}{Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress message about directory? DEFAULT: FALSE}

\item{read}{Read the file with \code{\link{readDWD}}?
If FALSE, only download is performed. DEFAULT: TRUE}

\item{format}{Format used in \code{\link{strptime}} to convert date/time column,
see \code{\link{readDWD}}. DEFAULT: NA}

\item{sleep}{If not 0, a random number of seconds between 0 and \code{sleep}
is passed to \code{\link{Sys.sleep}} after each download
to avoid getting kicked off the FTP-Server. DEFAULT: 0}

\item{progbar}{Logical: present a progress bar?
Only works if the R package pbapply is available. DEFAULT: TRUE}
}
\value{
presuming downloading and processing were successfull:
        if \code{read=TRUE}, a data.frame of the desired dataset
        (as returned by \code{\link{readDWD}}),
        otherwise the filenames as saved on disc
        (may have "_n" appended in name, see \code{\link{fileDWD}}).\cr
        if length(file)>1, the output is a list of data.frames / filenames.
}
\description{
Get climate data from the German Weather Service (DWD) FTP-server.
The desired .zip dataset is downloaded into \code{dir}.
If \code{read=TRUE}, it is also read, processed and returned as a data.frame
}
\examples{
# toDo: get from berryFunctions::dataDWD

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun-Oct 2016
}
\seealso{
\code{\link{readDWD}}, \code{\link{download.file}}.
         Helpful for plotting: \code{berryFunctions::\link[berryFunctions]{monthAxis}},
         see also \code{berryFunctions::\link[berryFunctions]{climateGraph}}
}
\keyword{data}
\keyword{file}

