% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{fileIndex}
\alias{fileIndex}
\title{Files available on the DWD CDC FTP server}
\format{data.frame with character strings. 25'631 rows x 7 columns:
        \code{res}, \code{var}, \code{per} (see \code{\link{selectDWD}}),
        station \code{id} and time series \code{start} and \code{end}
        according to \code{path}.}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
A data.frame with the filenames with path starting at the default \code{base} value:
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}.
Created with \code{\link{createIndex}} in the last section of
\url{https://github.com/brry/rdwd/blob/master/R/meta.R}
In functions, you can access it with \code{rdwd:::fileIndex}.
}
\examples{

data(fileIndex)
head(fileIndex)

# in functions, you can use head(rdwd:::fileIndex), but I don't export it
# because Hadley says 'Never @export a data set' in
# browseURL("http://r-pkgs.had.co.nz/data.html#data-data")

}
\seealso{
\code{\link{metaIndex}}, \code{\link{indexDWD}}, \code{\link{selectDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June-Nov 2016
}
\keyword{datasets}
