% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fread_s3}
\alias{fread_s3}
\title{fread data from s3 empty}
\usage{
fread_s3(station_id, ids = TRUE, zip = TRUE,
  s3_path = "s3://dwd-oekoproj/", loc_path = tempdir())
}
\arguments{
\item{station_id}{if ids = TRUE numerical vector of station_id's else character-vector of files to download from S3}

\item{ids}{logical see station_id}

\item{zip}{logical if the files/ids specified in station_id are ziped}

\item{s3_path}{character path to s3-Bucket}

\item{loc_path}{path where the data is saved temporarily}
}
\value{
List of data.tables which were downloaded
}
\description{
fread data from s3 empty
}
\examples{
fread_s3(183)
}

