% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{index}
\alias{index}
\alias{fileIndex}
\alias{metaIndex}
\alias{geoIndex}
\title{Indexes of files and metadata on the DWD CDC FTP server}
\format{\bold{fileIndex}: data.frame with character strings. ca 25k rows x 7 columns:\cr
        \code{res}, \code{var}, \code{per} (see \code{\link{selectDWD}}),
        station \code{id} and time series \code{start} and \code{end}
        according to \code{path}.\cr
\bold{metaIndex}: data.frame with ca 35k rows for 12 columns:\cr
        \code{Stations_id, von_datum, bis_datum,
        Stationshoehe, geoBreite, geoLaenge, Stationsname, Bundesland,
        res, var, per, hasfile} \cr
\bold{geoIndex}: data.frame with ca 6k rows for 11 columns:\cr
        \code{id, name, state, lat, lon, ele, nfiles, nonpublic, recentfile, display, col}}
\source{
Deutscher WetterDienst / Climate Data Center  FTP Server
}
\description{
Created with \code{\link{createIndex}} in the last section of
\url{https://github.com/brry/rdwd/blob/master/R/rdwd-package.R}
In functions, you can access them with \code{rdwd:::fileIndex} etc.\cr
\bold{fileIndex}: A data.frame with the filenames at the default \code{base} value
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}.\cr
\bold{metaIndex}: A data.frame with the contents of all the station description files
(..._Beschreibung_Stationen.txt) in the folders hourly, daily and monthly at \code{base}.\cr
\bold{geoIndex}: \code{metaIndex} distilled to geographic locations.
}
\examples{
data(fileIndex)
data(metaIndex)
data(geoIndex)
head(fileIndex)
head(metaIndex)
head(geoIndex)

# in functions, you can use head(rdwd:::fileIndex) etc, but I don't export them
# because Hadley says 'Never @export a data set' in
# browseURL("http://r-pkgs.had.co.nz/data.html#data-data")

}
\seealso{
\code{\link{createIndex}}, \code{\link{indexFTP}}, \code{\link{selectDWD}},
         \code{\link{findID}}, \code{\link{metaInfo}},
         \href{../doc/mapDWD.html}{\code{vignette("mapDWD")}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June-Nov 2016, June 2017
}
\keyword{datasets}
