% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2df.R
\name{index2df}
\alias{index2df}
\title{process the index of the DWD CDC FTP Server}
\usage{
index2df(path, dir = "DWDdata", filename = "INDEX.txt", quiet = FALSE)
}
\arguments{
\item{path}{Char: vector of paths returned by \code{\link{indexDWD}} called
with the default \code{base} value.}

\item{dir}{Char: writeable directory name where to save the output.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{filename}{Char: Name of file in \code{dir} in which to save the output.
Use \code{filename=""} to suppress writing. DEFAULT: "INDEX.txt"}

\item{quiet}{Logical: Suppress messages about directory / filename? DEFAULT: FALSE}
}
\value{
invisible dataframe with a number of columns tried to infer from the paths.
        This is also saved to disc.
}
\description{
Create a data.frame out of the vector index returned by \code{\link{indexDWD}}.
\code{index2df} tries to obtain res, var, time, file, id, start and end from the paths.
This produces the \code{\link{fileIndex}} used in \code{\link{selectDWD}}. ToDO: create and document 'metaIndex'
}
\examples{
# see last part of
if(interactive())
browseURL("https://github.com/brry/rdwd/blob/master/R/rdwd-package.R")

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{indexDWD}}, \code{\link{fileIndex}}, \code{\link{selectDWD}}
}
\keyword{manip}

