% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexDWD.R
\name{indexDWD}
\alias{indexDWD}
\title{Create a recursive index of the DWD CDC FTP Server}
\usage{
indexDWD(folder = "",
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  ziponly = FALSE, sleep = 0, dir = "DWDdata", quiet = FALSE,
  progbar = !quiet)
}
\arguments{
\item{folder}{Folder to be indexed recursively, e.g. "/hourly/wind/".
DEFAULT: "" (all folders at \code{base})}

\item{base}{Main directory of DWD ftp server, defaulting to observed climatic records.
DEFAULT: \url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}}

\item{ziponly}{Logical: return only the filenames ending in ".zip"?
DEFAULT: FALSE}

\item{sleep}{If not 0, a random number of seconds between 0 and \code{sleep}
is passed to \code{\link{Sys.sleep}} after each read folder
to avoid getting kicked off the FTP-Server. DEFAULT: 0}

\item{dir}{Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress message about directory and failed
\code{RCurl::\link[RCurl]{getURL}}? DEFAULT: FALSE}

\item{progbar}{Logical: present a progress bar in each level?
Only works if the R package pbapply is available. DEFAULT: TRUE}
}
\value{
currently a vector with file paths (output may change in the future)
}
\description{
Create a list of all the files (in subfolders) at the Climate Data Center (CDC)
FTP-Server from the German Weather Service (DWD, Deutscher WetterDienst) at
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}.\cr
The R package \code{RCurl} must be available to do this.
If \code{RCurl::\link[RCurl]{getURL}} fails, usually because bot access is
detected and denied, there will stil be an output which you can pass in a
second run via \code{folder} to extract the remaining dirs.
You might want to wait a bit and set \code{sleep} to a higher value in that case.
}
\details{
It's not suggested to run this for all folders, as it can take quite some time
and you may get kicked off the FTP-Server. This package contains an index
of the climatic observations at weather stations: \code{View(rdwd:::\link{fileIndex})}
If it is out of date, please let me know!
}
\examples{
\dontrun{ ## Needs internet connection
sol <- indexDWD(folder="/daily/solar")
head(sol)
rm(sol)


# Here's how I produce and update   ?fileIndex    ---------------------------
# index <- indexDWD(sleep=30) # commented out to prevent accidental calling
# index <- indexDWD(index, sleep=30) # potentially needed several times
index <- readLines("DWDdata/INDEX_of_DWD_.txt") # 25'631 elements (2016-10-21)
indexcompare <- index2df(index)
fileIndex <- read.table("DWDdata/INDEX.txt", sep="\\t", header=TRUE, colClasses="character")
stopifnot(all(fileIndex==indexcompare))

save(fileIndex, file="data/fileIndex.rda")
tools::resaveRdaFiles("data/fileIndex.rda")
#devtools::use_data(fileIndex, internal=TRUE)


# Here's how I produce and update   ?metaIndex    ----------------------------
sel <- substr(fileIndex$path, nchar(fileIndex$path)-3, 1e4)==".txt"
sel <- sel & grepl("Beschreibung", fileIndex$path)
fileIndex[sel, -(4:6)]
base <- "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate"

# comparison with all folders:
sum(sel) #27
folders <- with(fileIndex, unique(paste(res,var,time, sep="/"))   )
length(folders) # 30 (3 multi_annual, no Beschreibung file)
folders;  rm(folders)

# ToDo: actually create metalist

}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\keyword{file}

