% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{indexlist}
\alias{indexlist}
\title{Files available on the DWD CDC FTP server}
\format{data.frame with character srings. 25'631 rows x 7 columns:
        \code{res}, \code{var}, \code{time} (see \code{\link{metaDWD}}),
        station \code{id},
        \code{start} and \code{end} of time series according to \code{path}.}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
A data.frame with the filenames with path starting at the default \code{base} value:
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}.
Created with \code{\link{index2df}} in the not-tested example section of \code{\link{indexDWD}}.
}
\examples{

data(indexlist)
head(indexlist)

}
\keyword{datasets}

