% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localtestdir.R
\name{localtestdir}
\alias{localtestdir}
\title{local test data directory}
\usage{
localtestdir(packdir = ".",
  folder = "localtests/CreateVignettes/DWDdata", file = NULL)
}
\arguments{
\item{packdir}{Path to package directory. DEFAULT: "."}

\item{folder}{Path inside package. DEFAULT: "localtests/CreateVignettes/DWDdata"}

\item{file}{Optional: path(s) at \code{folder}. DEFAULT: NULL}
}
\value{
charstring (directory)
}
\description{
returns a directory used for local tests on Berry's computers.
This is used in many examples to reduce the number of downloads by saving 
the downloaded DWD data in this directory.
My local test script is at \url{https://github.com/brry/rdwd/blob/master/localtests/localtests.R}.
It runs all the examples, even the dontrun sections.
}
\examples{
localtestdir()

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr 2019
}
\keyword{file}
