% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{mapDWD}
\alias{mapDWD}
\title{interactive map of data available on the DWD CDC FTP server}
\format{leaflet map}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
leaflet map from \code{\link{geoIndex}}
}
\examples{

data(mapDWD)
library(leaflet)
mapDWD
# in functions, you can use rdwd:::mapDWD


}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2017
}
\seealso{
\code{\link{geoIndex}}, \code{\link{metaInfo}}
}
\keyword{datasets}

