% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDWD.R
\name{metaDWD}
\alias{metaDWD}
\title{Get metadata from the DWD CDC FTP Server}
\usage{
metaDWD(id = "", name = "",
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  res = "hourly", var = "precipitation", time = c("recent", "historical"),
  dir = "DWDdata", quiet = FALSE, browse = FALSE, files = FALSE,
  ziponly = TRUE)
}
\arguments{
\item{id, name}{Optional character strings. If !="" and corresponding
file(s) exist(s), metaDWD returns "base/res/var/time/filename.zip"
CURRENTLY IGNORED! ToDo: match with indexed metadata}

\item{base}{Main directory of DWD ftp server, defaulting to observed climatic records.
DEFAULT: ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate"}

\item{res}{Time resolution available at \code{base}. usually one of
hourly","daily","monthly". DEFAULT: "hourly"}

\item{var}{Weather variable of interest, e.g.   ToDo: put overview from index here
DEFAULT: "precipitation"}

\item{time}{Desired time range. One of
"recent" (data from the last year, up to date usually within a few days) or
"historical" (long time series). Can be abbreviated (partial matching).
DEFAULT: "recent" (the first of the inputs is used)}

\item{dir}{Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress messages about directory / filename? DEFAULT: FALSE}

\item{browse}{Logical: open repository via \code{\link{browseURL}}?
If TRUE, no metadata is downloaded, but the path URL is returned.
DEFAULT: FALSE}

\item{files}{Logical: instead of station metadata, return a list of the
actually available files? Will call code{\link{indexDWD}}.
DEFAULT: FALSE}

\item{ziponly}{Logical: If files=TRUE, only return the zip files, not the
description (Beschreibung.txt) files? DEFAULT: TRUE}
}
\value{
data.frame with weather station metadata
}
\description{
Get station metadata. Useful to select files for downloading with \code{\link{dataDWD}}.
Column widths for \code{\link{read.fwf}} are computed internally)
}
\examples{
# toDo

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{dataDWD}}, \code{\link{indexDWD}}, \code{\link{readDWD}}
}
\keyword{file}

