% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{metaIndex}
\alias{metaIndex}
\title{station info (meta data) available on the DWD CDC FTP server}
\format{data.frame with character srings. 36'254 rows x 11 columns:
        \code{Stations_id}, \code{von_datum}, \code{bis_datum}
        \code{Stationshoehe}, \code{geoBreite}, \code{geoLaenge}
        \code{Stationsname}, \code{Bundesland},
        \code{res}, \code{var}, \code{time} (see \code{\link{selectDWD}})}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
A data.frame with the contents of all the station description files
(..._Beschreibung_Stationen.txt) in the folders hourly, daily and monthly at
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}.
Created with \code{\link{index2df}} in the last section of
\url{https://github.com/brry/rdwd/blob/master/R/rdwd-package.R}
In functions, you can access it with \code{rdwd:::metaIndex}.
}
\examples{

data(metaIndex)
head(metaIndex)

# in functions, you can use head(rdwd:::metaIndex)

}
\seealso{
\code{\link{fileIndex}}, \code{\link{indexDWD}}
}
\keyword{datasets}

