% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVars.R
\docType{data}
\name{parameter_abbreviations}
\alias{parameter_abbreviations}
\title{Parameter abbreviations for data on the DWD CDC FTP server}
\format{An object of class \code{data.frame} with 118 rows and 2 columns.}
\usage{
parameter_abbreviations
}
\description{
Parameter abbreviations for data on the DWD CDC FTP server
}
\examples{
head(parameter_abbreviations)

}
\seealso{
\code{\link{readVars}}, \code{\link{readDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2018
}
\keyword{datasets}
