% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{package}
\name{rdwd}
\alias{rdwd}
\alias{rdwd-package}
\title{Download Climate Data from DWD (German Weather Service)}
\description{
Select weather data from the DWD (Deutscher Wetterdienst) with
\code{\link{selectDWD}}, which uses \code{\link{fileIndex}} and \code{\link{metaIndex}}.
Download and process data sets with \code{\link{dataDWD}} and \code{\link{readDWD}}.\cr
For an introduction to the package, see \url{https://github.com/brry/rdwd#rdwd}
}
\details{
The following folders are available (and a few more).
"<" signifies a split into the folders "recent" and "historical".
"-" signifies that there are no further sub-folders.
\tabular{lll}{
\bold{hourly}      \tab | \bold{daily}       \tab | \bold{monthly}\cr
air_temperature <  \tab | kl <               \tab | kl <          \cr
cloudiness <       \tab | more_precip <      \tab | more_precip < \cr
precipitation <    \tab |                    \tab |               \cr
pressure <         \tab |                    \tab |               \cr
sun <              \tab |                    \tab |               \cr
wind <             \tab |                    \tab |               \cr
soil_temperature < \tab | soil_temperature < \tab |               \cr
solar -            \tab | solar -            \tab |               \cr
}
}
\examples{

# ToDo

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2016
}
\seealso{
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}
}
\keyword{documentation}
\keyword{package}

