% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rdwdquiet}
\alias{rdwdquiet}
\title{global quiet option for rdwd}
\usage{
rdwdquiet()
}
\description{
global quiet option. The default \code{rdwdquiet()} is \code{FALSE}.\cr
Just write the following in your code and all subsequent calls will be quiet:\cr
\code{options(rdwdquiet=TRUE)}
}
