% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_dwd_dat}
\alias{read_dwd_dat}
\title{Read and save archives from DWD}
\usage{
read_dwd_dat(archive, write_station_dat, logging = TRUE, ...)
}
\arguments{
\item{archive}{data.frame with columns "url", "station_id", "historical", "type" see ?archives_dwd}

\item{write_station_dat}{function how to save the data}

\item{logging}{logical | If TRUE a log-file of the process is generated and saved as "read_dwd_dat_log.txt"}

\item{...}{further parameters for write_station_dat(dat, ...)}
}
\value{
data.frame | which station measures what
}
\description{
Read and save archives from DWD
}
\examples{
data(archives_dwd)
archives_dwd <- archives_dwd[which(archives_dwd$station_id == 183 & archives_dwd$type == "wind"),]
read_dwd_dat(archives_dwd, write_dwd_csv, logging = FALSE)
if(!file.exists("station183.csv")) error("station183.csv not found")
file.remove("station183.csv")
}

