% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.asczip}
\alias{readDWD.asczip}
\title{read dwd gridded asc data}
\usage{
readDWD.asczip(file, dividebyten, quiet = rdwdquiet(), ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g.
DWDdata/annual/radiation_global/grids_germany_annual_radiation_global_2024.zip}

\item{dividebyten}{Logical: Divide the numerical values by 10? See \link{readDWD}.
DEFAULT: TRUE}

\item{quiet}{Ignored.
DEFAULT: FALSE through \code{\link[=rdwdquiet]{rdwdquiet()}}}

\item{\dots}{Further arguments passed to \code{\link[terra:rast]{terra::rast()}}}
}
\value{
\code{\link[terra:rast]{terra::rast}} object
}
\description{
Read gridded asc data.
Intended to be called via \code{\link[=readDWD]{readDWD()}}.\cr
}
\examples{
\dontrun{ # Excluded from CRAN checks, but run in localtests
link <- "annual/radiation_global/grids_germany_annual_radiation_global_2023.zip" # 0.1 MB
grid <- dataDWD(link, base=gridbase, joinbf=TRUE)
plotRadar(grid, proj="seasonal", extent=NULL)
}
}
\seealso{
\code{\link[=readDWD]{readDWD()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2025
}
