% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.binary}
\alias{readDWD.binary}
\title{read dwd gridded radolan binary data}
\usage{
readDWD.binary(file, progbar = TRUE, selection = NULL, ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g. 
DWDdata/grids_germany_daily_radolan_historical_2017_SF201712.tar.gz}

\item{progbar}{Show messages and progress bars? \code{\link{readDWD}} will
keep progbar=TRUE for binary files, even if length(file)==1.
DEFAULT: TRUE}

\item{selection}{Optionally read only a subset of the ~24*31=744 files.
Called as \code{f[selection]}. DEFAULT: NULL (ignored)}

\item{\dots}{Further arguments passed to \code{\link{readBin}}}
}
\value{
vector
}
\description{
This does not work correctly yet for the tested data 
(grids_germany/daily/radolan/historical)! Hints are welcome!
}
\seealso{
\code{\link{readDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2018
}
