% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.meta}
\alias{readDWD.meta}
\title{read dwd metadata (Beschreibung*.txt files)}
\usage{
readDWD.meta(file, ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g. 
DWDdata/daily_kl_recent_KL_Tageswerte_Beschreibung_Stationen.txt}

\item{\dots}{Further arguments passed to \code{\link{read.fwf}}}
}
\value{
data.frame
}
\description{
read dwd metadata (Beschreibung*.txt files).
 Column widths for \code{\link{read.fwf}} are computed internally.\cr
 if(any(meta)), \code{\link{readDWD}} tries to set the locale to German 
 (to handle Umlaute correctly). It is hence not recommended to call
 \code{rdwd:::readDWD.meta} directly on a file!\cr
 Names can later be changed to ascii with 
 \code{berryFunctions::\link{convertUmlaut}}.
}
\seealso{
\code{\link{readDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}
}
