% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.multia}
\alias{readDWD.multia}
\title{read multi_annual dwd data}
\usage{
readDWD.multia(file, ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g. 
DWDdata/multi_annual_mean_81-10_Temperatur_1981-2010_aktStandort.txt or
DWDdata/multi_annual_mean_81-10_Temperatur_1981-2010_Stationsliste_aktStandort.txt}

\item{\dots}{Further arguments passed to \code{\link{read.table}}}
}
\value{
data.frame
}
\description{
All other observational data at \code{\link{dwdbase}} 
(\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}) can be read
with \code{\link{readDWD.data}}, except for the multi_annual data.
}
\seealso{
\code{\link{readDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2019
}
