% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.raster}
\alias{readDWD.raster}
\title{read dwd gridded raster data}
\usage{
readDWD.raster(file, dividebyten, ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g. 
DWDdata/grids_germany/seasonal/air_temperature_mean/
16_DJF_grids_germany_seasonal_air_temp_mean_188216.asc.gz}

\item{dividebyten}{Logical: Divide the numerical values by 10?
DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{raster::\link[raster]{raster}}}
}
\value{
\code{raster::\link[raster]{raster}} object
}
\description{
read dwd gridded raster data
}
\examples{
\dontrun{ # Excluded from CRAN checks, but run in localtests

rasterbase <- paste0(gridbase,"/seasonal/air_temperature_mean")
ftp.files <- indexFTP("/16_DJF", base=rasterbase, dir=tempdir())
localfiles <- dataDWD(ftp.files[1:2], base=rasterbase, joinbf=TRUE,
                      dir=localtestdir(), read=FALSE)
rf <- readDWD(localfiles[1])
rf <- readDWD(localfiles[1]) # needs to be able to run a second time
raster::plot(rf)

testthat::expect_equal(raster::cellStats(rf, range), c(-8.2,4.4))
rf10 <- readDWD(localfiles[1], dividebyten=FALSE)
raster::plot(rf10)
testthat::expect_equal(raster::cellStats(rf10, range), c(-82,44))
}
}
\seealso{
\code{\link{readDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2018
}
