% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.rklim}
\alias{readDWD.rklim}
\title{read dwd gridded radklim binary data}
\usage{
readDWD.rklim(
  file,
  exdir = NULL,
  unpacked = NULL,
  selection = NULL,
  toraster = TRUE,
  quiet = rdwdquiet(),
  progbar = !quiet,
  ...
)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g.
DWDdata/5_minutes_radolan_reproc_2017_002_bin_2020_YW2017.002_202006.tar}

\item{exdir}{Directory to unzip into. If existing, only the needed files
will be unpacked with \code{\link[=untar]{untar()}}. Note that exdir
size will be around 17 GB for 5-minute files.
If \code{unpacked=FALSE}, exdir can contain other files
that will be ignored for the actual reading.
DEFAULT: basename(file) at tempdir}

\item{unpacked}{Manually indicate whether .tar.gz files within .tar file
have already been unpacked before.
DEFAULT: NULL: checks if 'yw.*--bin' file(s) are present}

\item{selection}{Optionally read only a subset of the ~ 12 x 24 x 30/31 = 8640 files.
Called as \code{f[selection]}. DEFAULT: NULL (ignored)}

\item{toraster}{Logical: convert to raster stack? see \link{readDWD.binary}
DEFAULT: TRUE}

\item{quiet}{Suppress progress messages?
DEFAULT: FALSE through \code{\link[=rdwdquiet]{rdwdquiet()}}}

\item{progbar}{Show progress bars?
DEFAULT: !quiet, i.e. TRUE}

\item{\dots}{Further arguments passed to \code{\link[dwdradar:readRadarFile]{dwdradar::readRadarFile()}},
i.e. \code{na} and \code{clutter}}
}
\value{
list depending on argument \code{toraster}, see there for details
}
\description{
read gridded radklim binary data.
Intended to be called via \code{\link[=readDWD]{readDWD()}}.\cr
Note: needs dwdradar >= 0.2.6 (2021-08-08)
}
\examples{
\dontrun{ # Excluded from CRAN checks, but run in localtests
yw_link <- "/5_minutes/radolan/reproc/2017_002/bin/2020/YW2017.002_202006.tar"
yw_file <- dataDWD(url=yw_link, base=gridbase, joinbf=TRUE, dir=locdir(), read=FALSE)
x <- readDWD(yw_file, selection=3641:3644)
raster::plot(x$dat)

f <- system.file("tests//raa01-yw2017.002_10000-2006131525-dwd---bin", package="dwdradar")
x <- dwdradar::readRadarFile(f)
x$dat <- raster::raster(x$dat)
raster::plot(x$dat)
plotRadar(x$dat)
}
}
\seealso{
\code{\link[=readDWD.binary]{readDWD.binary()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2021.
}
