% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\name{release_questions}
\alias{release_questions}
\title{Reminders when using devtools::release}
\usage{
release_questions()
}
\description{
Reminders when using devtools::release. Code is in R/rdwd-package.R
}
\keyword{internal}
