% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{rowDisplay}
\alias{rowDisplay}
\title{create display character string for leaflet map popup from data.frame rows}
\usage{
rowDisplay(x)
}
\arguments{
\item{x}{data.frame with colnames}
}
\value{
Vector of characterstrings, one for each row in x.
}
\description{
create display character string for leaflet map popup from data.frame rows
}
\seealso{
\code{\link{geoIndex}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{character}
