% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRdwd.R
\name{updateRdwd}
\alias{updateRdwd}
\title{Update rdwd development version}
\usage{
updateRdwd(
  pack = "rdwd",
  user = "brry",
  vignette = TRUE,
  quiet = rdwdquiet(),
  ...
)
}
\arguments{
\item{pack}{Name of (already installed) package. DEFAULT: "rdwd"}

\item{user}{Github username. repo will then be user/pack. DEFAULT: "brry"}

\item{vignette}{build_vignettes in \code{\link[remotes:install_github]{remotes::install_github()}}?
DEFAULT: TRUE}

\item{quiet}{Suppress version messages and \code{remotes::install} output?
DEFAULT: FALSE through \code{\link[=rdwdquiet]{rdwdquiet()}}}

\item{\dots}{Further arguments passed to \code{\link[remotes:install_github]{remotes::install_github()}}}
}
\value{
data.frame with version information
}
\description{
Update rdwd to the latest development version on github, if necessary.
If the version number or date is larger on github,
\code{\link[remotes:install_github]{remotes::install_github()}} will be called.
}
\examples{
# updateRdwd()

}
\seealso{
\code{\link[remotes:install_github]{remotes::install_github()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2019
}
\keyword{file}
