% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileType.R
\docType{data}
\name{validFileTypes}
\alias{validFileTypes}
\title{valid fileType values}
\format{
An object of class \code{character} of length 13.
}
\usage{
validFileTypes
}
\description{
fileType values that have a reading subfunction \code{readDWD.ftype()}.
}
\keyword{datasets}
