% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\name{viewIndex}
\alias{viewIndex}
\title{View fileIndex and metaIndex}
\usage{
viewIndex()
}
\description{
Open \code{rdwd:::\link{fileIndex}} and \code{rdwd:::\link{metaIndex}} with \code{\link{View}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}

