% Generated by roxygen2 (4.0.2): do not edit by hand
\name{write_dwd_s3}
\alias{write_dwd_s3}
\title{Saves a data.table as csv- or tar.gz file and moves it to S3}
\usage{
write_dwd_s3(dat, zip = TRUE, dir_s3 = stop("plese privide s3 directory"))
}
\arguments{
\item{dat}{a data.table}

\item{zip}{logical}

\item{dir}{character string of S3 bucket/directory like "my_bucket"}
}
\value{
Nothing
}
\description{
Saves a data.table as csv- or tar.gz file and moves it to S3
}

