% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBorders.R
\name{addBorders}
\alias{addBorders}
\title{add country and Bundesland borders to a map}
\usage{
addBorders(de = "grey80", eu = "black", add = TRUE, ...)
}
\arguments{
\item{de}{Color for Bandeslaender line. NA to suppress. DEFAULT: "grey80"}

\item{eu}{Color for countries line. NA to suppress. DEFAULT: "black"}

\item{add}{Logical: add to existing plot? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{raster::\link[raster]{plot}}}
}
\value{
invisible list with DEU and EUR
}
\description{
add country and Bundesland borders to a map
}
\examples{
if(requireNamespace("raster", quietly=TRUE)){
plot(1, xlim=c(2,16), ylim=c(47,55)) 
addBorders()
plot(1, xlim=c(2,16), ylim=c(47,55))
addBorders(de="orange", eu=NA)
}

}
\seealso{
\code{\link{DEU}}, \code{\link{EUR}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2019
}
\keyword{aplot}
