% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDWD.R
\name{dataDWD}
\alias{dataDWD}
\title{Download data from the DWD CDC FTP Server}
\usage{
dataDWD(file, dir = "DWDdata", progbar = TRUE, sleep = 0, quiet = FALSE,
  meta = substr(file, nchar(file) - 3, 10000) == ".txt", read = TRUE,
  format = NA, browse = FALSE, ...)
}
\arguments{
\item{file}{Char: complete file URL (including base and filename.zip) as returned by
\code{\link{selectDWD}}. Can also be a vector with several filenames.}

\item{dir}{Single char: Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{progbar}{Single logical: present a progress bar?
Only works if the R package \code{pbapply} is available. DEFAULT: TRUE}

\item{sleep}{Single number. If not 0, a random number of seconds between 0 and
\code{sleep} is passed to \code{\link{Sys.sleep}} after each download
to avoid getting kicked off the FTP-Server. DEFAULT: 0}

\item{quiet}{Logical: suppress message about directory / filenames? DEFAULT: FALSE}

\item{meta}{Logical: is the \code{file} a meta file?
DEFAULT: TRUE for each file ending in ".txt"}

\item{read}{Logical: read the file with \code{\link{readDWD}}? If FALSE,
only download is performed and the filename returned. DEFAULT: TRUE}

\item{format}{Char: format used in \code{\link{strptime}} to convert date/time column,
see \code{\link{readDWD}}. DEFAULT: NA}

\item{browse}{Logical: open repository via \code{\link{browseURL}}? If TRUE,
no data is downloaded, but the URL path without filename is returned.
DEFAULT: FALSE}

\item{\dots}{Further arguments passed to code{\link{download.file}}}
}
\value{
Presuming downloading and processing were successfull:
        if \code{read=TRUE}, a data.frame of the desired dataset
        (as returned by \code{\link{readDWD}}),
        otherwise the filenames as saved on disc
        (may have "_n" appended in name, see \code{\link{fileDWD}}).\cr
        If length(file)>1, the output is a list of data.frames / filenames.\cr
        The output is always invisible.
}
\description{
Get climate data from the German Weather Service (DWD) FTP-server.
The desired .zip (or .txt) dataset is downloaded into \code{dir}.
If \code{read=TRUE}, it is also read, processed and returned as a data.frame.
All arguments (except for \code{dir}, \code{progbar} and \code{sleep})
can be a vecor and will be recycled to the length of \code{file}.
}
\examples{
# toDo: get from berryFunctions::dataDWD
# d <- dataDWD(selectDWD(id="05692", res="daily", var="kl", time="recent"))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun-Oct 2016
}
\seealso{
\code{\link{selectDWD}}. \code{\link{readDWD}}, \code{\link{download.file}}.
         Helpful for plotting: \code{berryFunctions::\link[berryFunctions]{monthAxis}},
         see also \code{berryFunctions::\link[berryFunctions]{climateGraph}}
}
\keyword{data}
\keyword{file}

