% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDWD.R
\name{dataDWD}
\alias{dataDWD}
\title{Download data from the DWD CDC FTP Server}
\usage{
dataDWD(file, dir = "DWDdata", sleep = 0, quiet = FALSE,
  progbar = !quiet, browse = FALSE, read = TRUE, meta = substr(file,
  nchar(file) - 3, 10000) == ".txt", format = NA, ...)
}
\arguments{
\item{file}{Char (vector): complete file URL(s) (including base and filename.zip) as returned by
\code{\link{selectDWD}}. Can be a vector with several filenames.}

\item{dir}{Char: Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{sleep}{Number. If not 0, a random number of seconds between 0 and
\code{sleep} is passed to \code{\link{Sys.sleep}} after each download
to avoid getting kicked off the FTP-Server. DEFAULT: 0}

\item{quiet}{Logical: suppress message about directory / filenames? DEFAULT: FALSE}

\item{progbar}{Logical: present a progress bar with estimated remaining time?
If missing and length(file)==1, progbar is internally set to FALSE.
Only works if the R package \code{pbapply} is available. DEFAULT: TRUE (!quiet)}

\item{browse}{Logical: open repository via \code{\link{browseURL}} and
return URL folder path? If TRUE, no data is downloaded.
If file has several values, only unique folders will be opened.
DEFAULT: FALSE}

\item{read}{Logical: read the file(s) with \code{\link{readDWD}}? If FALSE,
only download is performed and the filename(s) returned. DEFAULT: TRUE}

\item{meta}{Logical (vector): is the \code{file} a meta file? Passed to
\code{\link{readDWD}}. DEFAULT: TRUE for each file ending in ".txt"}

\item{format}{Char (vector): format used in \code{\link{strptime}} to convert date/time column,
see \code{\link{readDWD}}. DEFAULT: NA}

\item{\dots}{Further arguments passed to code{\link{download.file}}}
}
\value{
Presuming downloading and processing were successfull:
        if \code{read=TRUE}, a data.frame of the desired dataset
        (as returned by \code{\link{readDWD}}),
        otherwise the filename as saved on disc
        (may have "_n" appended in name, see \code{\link{fileDWD}}).\cr
        If length(file)>1, the output is a list of data.frames / vector of filenames.\cr
        The output is always invisible.
}
\description{
Get climate data from the German Weather Service (DWD) FTP-server.
The desired .zip (or .txt) dataset is downloaded into \code{dir}.
If \code{read=TRUE}, it is also read, processed and returned as a data.frame.
}
\examples{
# toDo: get from berryFunctions::dataDWD
# d <- dataDWD(selectDWD(id="05692", res="daily", var="kl", time="recent"))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun-Oct 2016
}
\seealso{
\code{\link{selectDWD}}. \code{\link{readDWD}}, \code{\link{download.file}}.
         Helpful for plotting: \code{berryFunctions::\link[berryFunctions]{monthAxis}},
         see also \code{berryFunctions::\link[berryFunctions]{climateGraph}}
}
\keyword{data}
\keyword{file}

