% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBorders.R
\docType{data}
\name{DEU}
\alias{DEU}
\title{Map of German states (Bundeslaender) from GADM through the \code{raster} package}
\format{
Formal class 'SpatialPolygons' [package "sp"] with 4 slots
}
\description{
Map of German states (Bundeslaender) from GADM through the \code{raster} package
}
\details{
Obtained with the code: \cr
\code{DEU1 <- raster::getData("GADM", country="DEU", level=1)}\cr
\code{DEU <- rgeos::gSimplify(DEU1, tol=0.02, topologyPreserve=FALSE)}\cr
\code{raster::plot(DEU1)}\cr
\code{raster::plot(DEU)}\cr
\code{save(DEU,        file="inst/extdata/DEU.rda")}\cr
\code{tools::resaveRdaFiles("inst/extdata/DEU.rda")}\cr
}
\seealso{
\code{\link{addBorders}}, \code{\link{EUR}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, May 2018
}
\keyword{datasets}
