% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirDWD.R
\name{dirDWD}
\alias{dirDWD}
\title{directory management for rdwd}
\usage{
dirDWD(dir = "DWDdata", quiet = FALSE)
}
\arguments{
\item{dir}{Writeable directory. Created if not existent.
DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress message about directory? DEFAULT: FALSE}
}
\value{
prior working directory
}
\description{
Manage directories in the rdwd package
}
\examples{
# see source code of dataDWD and metaDWD

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{dataDWD}}
}
\keyword{file}

