% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirDWD.R
\name{dirDWD}
\alias{dirDWD}
\title{directory management for rdwd}
\usage{
dirDWD(dir = "DWDdata", quiet = rdwdquiet())
}
\arguments{
\item{dir}{Char for dirDWD: writeable directory name. Created if not existent.
DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Logical: Suppress messages about creating dir? DEFAULT: FALSE through \code{\link{rdwdquiet}()}}
}
\value{
dirDWD invisibly returns the prior working directory as per \code{\link{setwd}}.
}
\description{
Manage directories with useful messages in the rdwd package.
}
\examples{
# see source code of dataDWD and metaDWD

}
\seealso{
\code{\link{dataDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\keyword{file}
