% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{dwdbase}
\alias{dwdbase}
\alias{gridbase}
\title{DWD FTP Server base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
dwdbase
}
\description{
base URLs to the DWD FTP Server\cr\cr
\strong{\code{dwdbase}}: observed climatic records at ftp:// variant of\cr
\url{https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/}\cr
See \href{https://brry.github.io/rdwd/available-datasets.html}{overview of available datasets}
and \href{https://brry.github.io/rdwd/station-selection.html}{usage suggestions}.\cr\cr\cr
\strong{\code{gridbase}}: spatially interpolated gridded data at ftp:// variant of\cr
\url{https://opendata.dwd.de/climate_environment/CDC/grids_germany/}\cr
See \href{https://brry.github.io/rdwd/raster-data.html}{usage suggestions}
}
\keyword{datasets}
