% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{dwdbase}
\alias{dwdbase}
\title{DWD FTP Server base URL}
\format{An object of class \code{character} of length 1.}
\usage{
dwdbase
}
\description{
base URL to DWD FTP Server / observed climatic records.\cr
    \url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}
}
\keyword{datasets}
