% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{dwdbase}
\alias{dwdbase}
\alias{gridbase}
\title{DWD FTP Server base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
dwdbase
}
\description{
base URLs to the DWD FTP Server\cr\cr
\strong{\code{dwdbase}}: observed climatic records at\cr
\url{ftp://opendata.dwd.de/climate_environment/CDC/observations_germany/climate}\cr
An overview of available datasets and usage suggestions can be found at\cr
\url{https://bookdown.org/brry/rdwd/available-datasets.html}\cr
\url{https://bookdown.org/brry/rdwd/station-selection.html}\cr\cr\cr
\strong{\code{gridbase}}: spatially interpolated gridded data at\cr
\url{ftp://opendata.dwd.de/climate_environment/CDC/grids_germany}\cr
Usage instructions can be found at\cr
\url{https://bookdown.org/brry/rdwd/raster-data.html}
}
\keyword{datasets}
