% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{dwdbase}
\alias{dwdbase}
\alias{gridbase}
\title{DWD FTP Server base URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
dwdbase
}
\description{
base URL to DWD FTP Server\cr\cr
\code{dwdbase}: observed climatic records at\cr
\url{ftp://opendata.dwd.de/climate_environment/CDC/observations_germany/climate}\cr\cr
\code{gridbase}: spatially interpolated gridded data at\cr
\url{ftp://opendata.dwd.de/climate_environment/CDC/grids_germany}
}
\keyword{datasets}
