% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBorders.R
\docType{data}
\name{EUR}
\alias{EUR}
\title{Map of Western European countries through the \code{rworldmap} package}
\format{
SpatialPolygonsDataFrame (package "sp") with 32 rows
}
\description{
Map of Western European countries through the \code{rworldmap} package
}
\details{
Use directly with:\cr
\code{load(system.file("extdata/EUR.rda", package="rdwd"))}\cr\cr
Obtained with the code: \cr\preformatted{EUR <- rworldmap::getMap("low")
EUR <- raster::crop(EUR, c(-11,25, 40,60))
raster::crs(EUR) <- raster::crs(DEU)
raster::plot(EUR)
save(EUR,        file="inst/extdata/EUR.rda", version=2)
tools::resaveRdaFiles("inst/extdata/EUR.rda", version=2)
}
}
\seealso{
\code{\link{addBorders}}, \code{\link{DEU}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2019
}
\keyword{datasets}
