% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBorders.R
\docType{data}
\name{EUR}
\alias{EUR}
\title{Map of Western European countries through the \code{rworldmap} package}
\format{
SpatialPolygonsDataFrame [package "sp"] with 32 rows
}
\description{
Map of Western European countries through the \code{rworldmap} package
}
\details{
Obtained with the code: \cr
\code{EUR <- rworldmap::getMap("low")}\cr
\code{EUR <- raster::crop(EUR, c(-5,20, 40,60)) }\cr
\code{raster::plot(EUR)}\cr
\code{save(EUR,        file="inst/extdata/EUR.rda", version=2)}\cr
\code{tools::resaveRdaFiles("inst/extdata/EUR.rda", version=2)}\cr
}
\seealso{
\code{\link{addBorders}}, \code{\link{DEU}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2019
}
\keyword{datasets}
