% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{geoIndex}
\alias{geoIndex}
\title{coordinatewise station info (meta data) available on the DWD CDC FTP server}
\format{data.frame with ca 7k rows for 9 columns:
        \code{id}, \code{name}, \code{state}
        \code{lat}, \code{long}, \code{ele}
        \code{all_elev}, \code{nfiles_coord}, \code{nfiles_id}}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
\code{\link{metaIndex}} distilled to geographic locations.
}
\examples{

data(geoIndex)
head(geoIndex)
# in functions, you can use head(rdwd:::geoIndex)

# Map of all precipitation stations:
if(FALSE){ ## requires extra package
if(!requireNameSpace("OSMscale")) install.packages("OSMscale")
library("OSMscale")
map <- pointsMap(lat, long, data=geoIndex, fx=0.06, fy=0.06)
pdf("DWDdata/RainfallStationsMap_nfiles.pdf", width=5)
plot(map)
scaleBar(map, x=0.05, y=0.03, abslen=200)
geoIndex <- sortDF(geoIndex, "nfiles_coord", decreasing=FALSE)
pp <- projectPoints(lat, long, data=geoIndex, to=posm())
points(pp, cex=0.6)
colPoints(pp$x, pp$y, geoIndex$nfiles_coord, cex=0.6, zlab="")
title(main="DWD stations: number of files on ftp server", line=3)
dev.off()
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June-Nov 2016
}
\seealso{
\code{\link{metaIndex}}, \code{\link{createIndex}}
}
\keyword{datasets}

