% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{geoIndex}
\alias{geoIndex}
\title{coordinatewise station info (meta data) available on the DWD CDC FTP server}
\format{data.frame with ca 7k rows for 9 columns:
        \code{id}, \code{name}, \code{state}
        \code{lat}, \code{long}, \code{ele}
        \code{all_elev}, \code{nfiles_coord}, \code{nfiles_id}}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
\code{\link{metaIndex}} distilled to geographic locations.
}
\examples{

data(geoIndex)
head(geoIndex)
# in functions, you can use head(rdwd:::geoIndex)

# example usages are in ?rdwd

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June-Nov 2016
}
\seealso{
\code{\link{metaIndex}}, \code{\link{createIndex}}
}
\keyword{datasets}

