% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index2df.R
\name{index2df}
\alias{index2df}
\title{process the index of the DWD CDC FTP Server}
\usage{
index2df(path, dir = "DWDdata", quiet = FALSE)
}
\arguments{
\item{path}{Vector of paths returned by \code{\link{indexDWD}} called with
the default \code{base} value.}

\item{dir}{Writeable directory name where to save the output.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress messages about directory / filename? DEFAULT: FALSE}
}
\value{
invisible dataframe with a number of columns tried to infer from the paths.
        This is also saved to disc.
}
\description{
Create a data.frame out of the vector index returned by \code{\link{indexDWD}}.
\code{index2df} tries to obtain res, var, time, file, id, start and end from the paths.
This produces the \code{\link{indexlist}} used in \code{\link{selectDWD}}. ToDO: create and document 'metaIndex'
}
\examples{
# see indexDWD

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{indexDWD}}, \code{\link{indexlist}}, \code{\link{selectDWD}}
}
\keyword{manip}

