% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexDWD.R
\name{indexDWD}
\alias{indexDWD}
\title{Create a recursive index of the DWD CDC FTP Server}
\usage{
indexDWD(folder = "",
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  ziponly = FALSE, sleep = 0, dir = "DWDdata", quiet = FALSE,
  progbar = TRUE)
}
\arguments{
\item{folder}{Folder to be indexed recursively, e.g. "/hourly/wind/".
DEFAULT: "" (all folders at \code{base})}

\item{base}{Main directory of DWD ftp server, defaulting to observed climatic records.
DEFAULT: ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate"}

\item{ziponly}{Logical: return only the filenames ending in ".zip"?
DEFAULT: FALSE}

\item{sleep}{If not 0, a random number of seconds between 0 and \code{sleep}
is passed to \code{\link{Sys.sleep}} after each read folder
to avoid getting kicked off the FTP-Server. DEFAULT: 0}

\item{dir}{Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress message about directory and failed
\code{RCurl::\link[RCurl]{getURL}}? DEFAULT: FALSE}

\item{progbar}{Logical: present a progress bar in each level?
Only works if the R package pbapply is available. DEFAULT: TRUE}
}
\value{
currently a vector with file paths (output may change in the future)
}
\description{
Create a list of all the files (in subfolders) at the Climate Data Center (CDC)
FTP-Server from the German Weather Service (DWD, Deutscher WetterDienst) at
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}.\cr
If it fails, because bot access is detected and denied, there will stil be an
output which you can pass in a second run via \code{folder} to extract the remaining dirs.
You might want to set \code{sleep=10} or similar in that case.
}
\details{
It's not suggested to run this for all folders, as it can take quite some time
and you may get kicked off the FTP-Server. This package contains an index
of the climatic observations at weather stations: \code{\link{indexlist}}
If it is out of date, please let me know!
}
\examples{
\dontrun{ ## Needs internet connection
sol <- indexDWD(folder="/daily/solar")
head(sol)
rm(sol)

# Here's how I produce and update   ?indexlist
# index1 <- indexDWD(sleep=30) # commented out to prevent accidental calling
# index1 <- indexDWD(index1, sleep=30) # potentially needed several times
index1 <- readLines("DWDdata/INDEX_of_DWD_.txt") # 25'631 elements (2016-10-21)
indexcompare <- index2df(index1)
indexlist <- read.table("DWDdata/INDEX.txt", sep="\\t", header=TRUE, colClasses="character")
stopifnot(all(indexlist==indexcompare))
save(indexlist, file="data/indexlist.rda")
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\keyword{file}

