% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexDWD.R
\name{indexDWD}
\alias{indexDWD}
\title{Create an index of the DWD CDC FTP Server}
\usage{
indexDWD(folder = "",
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  sleep = 0, dir = "DWDdata", quiet = FALSE)
}
\arguments{
\item{folder}{Folder to be indexed recursively, e.g. "hourly/wind/".
DEFAULT: "" (all folders at \code{base})}

\item{base}{Main directory of DWD ftp server, defaulting to observed climatic records.
DEFAULT: ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate"}

\item{sleep}{If not 0, a random number of seconds between 0 and \code{sleep}
is passed to \code{\link{Sys.sleep}} after each folder level
to avoid getting kicked off the FTP-Server. DEFAULT: 0}

\item{dir}{Writeable directory name where to save the downloaded file.
Created if not existent. DEFAULT: "DWDdata" at current \code{\link{getwd}()}}

\item{quiet}{Suppress message about directory? DEFAULT: FALSE}
}
\value{
currently a vector with file paths (output may change in the future)
}
\description{
Create a list of all the files (in subfolders) at the Climate Data Center (CDC)
FTP-Server from the German Weather Service (DWD, Deutscher WetterDienst) at
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}
}
\details{
It's not suggested to run this for all folders, as it can take quite some time
and you may get kicked off the FTP-Server. This package contains an index
of the 'regular' climatic observations at weather stations.
ToDo: put reference here.
If it is out of date, please let me know!
}
\examples{
\dontrun{ ## Needs internet connection
sol <- indexDWD(folder="daily/solar")
head(sol)
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\keyword{file}

