% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{data}
\name{indexlist}
\alias{indexlist}
\title{Files available on the DWD CDC FTP server}
\format{data.frame with character srings. 25'631 rows x 7 columns:
        \code{res}, \code{var}, \code{time} (see \code{\link{metaDWD}}),
        station \code{id} and time series \code{start} and \code{end}
        according to \code{path}.}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
A data.frame with the filenames with path starting at the default \code{base} value:
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}.
Created with \code{\link{index2df}} in the not-tested example section of \code{\link{indexDWD}}.
In functions, you can access it with \code{rdwd:::indexlist}.
}
\examples{

data(indexlist)
head(indexlist)

# in functions, you can use head(rdwd:::indexlist), but I don't export it
# because Hadley says 'Never @export a data set' in
# browseURL("http://r-pkgs.had.co.nz/data.html#data-data")

}
\keyword{datasets}

