% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldist.R
\name{lldist}
\alias{lldist}
\title{distance between lat-long coordinates}
\usage{
lldist(lat, long, data, r = 6371, i = 1L, trace = TRUE)
}
\arguments{
\item{lat, long}{Latitude (North/South) and longitude (East/West) coordinates in decimal degrees}

\item{data}{Optional: data.frame with the columns \code{lat} and \code{long}}

\item{r}{radius of the earth. Could be given in miles. DEFAULT: 6371 (km)}

\item{i}{Integer: Index element against which all coordinate pairs
are computed. DEFAULT: 1}

\item{trace}{Logical: trace the coordinate check with \code{\link{checkLL}}?
Should be set to FALSE in a \link{do.call} setting to avoid overhead
computing time. DEFAULT: TRUE}
}
\value{
Vector with distance(s) in km (or units of \code{r}, if \code{r} is changed)
}
\description{
Great-circle distance between points at lat-long coordinates.
Mostly a copy of OSMscale::earthDist Version 0.4.5 (2017-03-02).
Copied manually to avoid dependency hell.
Does not check coordinates. Not exported.
}
\examples{
d <- read.table(header=TRUE, sep=",", text="
lat, long
52.514687,  13.350012   # Berlin
51.503162,  -0.131082   # London
35.685024, 139.753365") # Tokio
lldist(lat, long, d)      # from Berlin to L and T: 928 and 8922 km
lldist(lat, long, d, i=2) # from London to B and T: 928 and 9562 km


}
\seealso{
\code{\link{degree}} for pre-formatting,
         \url{http://www.movable-type.co.uk/scripts/latlong.html}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2016 + Jan 2017.
        Angle formula from Diercke Weltatlas 1996, Page 245
}
\keyword{spatial}
