% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localtestdir.R
\name{localtestdir}
\alias{localtestdir}
\title{local test data directory}
\usage{
localtestdir(packdir = ".", folder = "misc/localdata", file = NULL)
}
\arguments{
\item{packdir}{Path to package directory. DEFAULT: "."}

\item{folder}{Path inside package. DEFAULT: "misc/localdata"}

\item{file}{Optional: path(s) at \code{folder}. DEFAULT: NULL}
}
\value{
charstring (directory)
}
\description{
returns a directory used for local tests on Berry's computers.
This is used in many examples to save the downloaded DWD data in this directory,
thus avoiding multiple downloads of the same file.
}
\examples{
localtestdir()

}
\seealso{
\code{\link[=runLocalTests]{runLocalTests()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr 2019
}
\keyword{file}
