% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\docType{data}
\name{mapDWD}
\alias{mapDWD}
\title{Interactive map of data available on the DWD CDC FTP server}
\format{leaflet map}
\source{
Deutscher WetterDienst / Climata Data Center  FTP Server
}
\description{
Interactive leaflet map with DWD weather stations.
Created from \code{\link{geoIndex}} in the last section of
\url{https://github.com/brry/rdwd/blob/master/R/meta.R}
}
\examples{

data(mapDWD)
library(leaflet)
mapDWD
# in functions, you can use rdwd:::mapDWD

# The first line when clicked on a point can be copied to R for more information

}
\seealso{
\code{\link{geoIndex}}, \code{\link{metaInfo}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jan 2017
}
\keyword{datasets}
