% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaInfo.R
\name{metaInfo}
\alias{metaInfo}
\title{Information for a station ID on the DWD CDC FTP server}
\usage{
metaInfo(id, hasfileonly = TRUE)
}
\arguments{
\item{id}{Station ID (integer number or convertible to one)}

\item{hasfileonly}{Logical: Only show entries that have files? DEFAULT: TRUE}
}
\value{
invisible data.frame. Also \code{\link{print}s} the output nicely formatted.
}
\description{
Information for a station ID on the DWD CDC FTP server
}
\examples{
metaInfo(2849)

}
\seealso{
\code{\link{metaIndex}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Nov 2016
}
\keyword{datasets}
