% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{package}
\name{rdwd}
\alias{rdwd}
\alias{rdwd-package}
\alias{rdwd-package}
\title{Download Climate Data from DWD (German Weather Service)}
\description{
Select weather data from the DWD (Deutscher Wetterdienst) with
\code{\link{selectDWD}} or \code{\link{nearbyStations}}. \cr
Download and process data sets with \code{\link{dataDWD}} and \code{\link{readDWD}}.\cr
Station selection is done offline with \code{\link{fileIndex}} and
\code{\link{findID}} (which uses \code{\link{metaIndex}}).\cr
The Index objects are created with \code{\link{indexDWD}} and \code{\link{createIndex}}.\cr
For an introduction to the package, see the \href{../doc/rdwd.html}{main vignette}.
}
\section{Searchability Terms}{

Weather Data Germany download with R, Climate Data Germany\cr
Deutscher Wetterdienst R Daten download Klimastationen\cr
DWD Daten mit R runterladen, Wetter und Klimadaten in R\cr
}

\seealso{
USA data: \url{https://www.rdocumentation.org/packages/countyweather},
         \url{https://www.rdocumentation.org/packages/rnoaa}\cr
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June-Nov 2016
}
\keyword{documentation}
\keyword{package}
