% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{package}
\name{rdwd}
\alias{rdwd}
\alias{rdwd-package}
\title{Download Climate Data from DWD (German Weather Service)}
\description{
Select weather data from the DWD (Deutscher Wetterdienst) with
\code{\link{metaDWD}}, \code{\link{indexDWD}} and \code{\link{indexlist}}.\cr
Download and process data sets with \code{\link{dataDWD}} and \code{\link{readDWD}}.
}
\details{
ToDo: more info
}
\examples{

# ToDo

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, 2016
}
\seealso{
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}
}
\keyword{documentation}
\keyword{package}

