% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{package}
\name{rdwd}
\alias{rdwd}
\alias{rdwd-package}
\title{Download Climate Data from DWD (German Weather Service)}
\description{
Select weather data from the DWD (Deutscher Wetterdienst) with
\code{\link{selectDWD}}, which uses \code{\link{fileIndex}} and
\code{\link{findID}} (which uses \code{\link{metaIndex}}).
The Index objects are created with \code{\link{indexDWD}} and \code{\link{index2df}}.\cr
Download and process data sets with \code{\link{dataDWD}} and \code{\link{readDWD}}.
code{\link{dirDWD}} and code{\link{fileDWD}} make sure no file is overwritten and give useful messages.\cr
For an introduction to the package, see \url{https://github.com/brry/rdwd#rdwd}
}
\details{
The following folders are available (and a few more at the \code{res} level) at
\url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}\cr
"<" signifies a split into the folders \code{time} = "recent" and "historical".\cr
"-" signifies that there are no further sub-folders.
\tabular{lll}{
\code{res}=\bold{hourly} \tab | \code{res}=\bold{daily} \tab | \code{res}=\bold{monthly} \cr
\code{var=} \tab \tab \cr
air_temperature <  \tab | kl <               \tab | kl <          \cr
cloudiness <       \tab | more_precip <      \tab | more_precip < \cr
precipitation <    \tab |                    \tab |               \cr
pressure <         \tab |                    \tab |               \cr
sun <              \tab |                    \tab |               \cr
wind <             \tab |                    \tab |               \cr
soil_temperature < \tab | soil_temperature < \tab |               \cr
solar -            \tab | solar -            \tab |               \cr
}
}
\examples{

# see README.md

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June-Oct 2016
}
\keyword{documentation}
\keyword{package}

