% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\docType{package}
\name{rdwd}
\alias{rdwd}
\alias{rdwd-package}
\title{Handle Climate Data from DWD (German Weather Service)}
\description{
\itemize{
\item find, select, download + read data from the German weather service DWD\cr
\item vectorized, progress bars, no re-downloads\cr
\item index of files + meta data\cr
\item observational time series from 6k meteorological recording stations (2.5k active)\cr
-> rain, temperature, wind, sunshine, pressure, cloudiness, humidity, snow, ...\cr
\item gridded raster data from radar + interpolation\cr
\item european data stock slowly growing\cr
For an introduction to the package, see \url{https://bookdown.org/brry/rdwd}.\cr
}
}
\section{Searchability Terms}{

Weather Data Germany download with R, Climate Data Germany\cr
Deutscher Wetterdienst R Daten download Klimastationen\cr
DWD Daten mit R runterladen, Wetter und Klimadaten in R
}

\seealso{
related weather data processing references at \url{https://bookdown.org/brry/rdwd/see-also.html}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}
}
\keyword{documentation}
\keyword{package}
