% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.binary}
\alias{readDWD.binary}
\title{read dwd gridded radolan binary data}
\usage{
readDWD.binary(file, progbar = TRUE, selection = NULL, ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g. 
DWDdata/grids_germany_daily_radolan_historical_2017_SF201712.tar.gz}

\item{progbar}{Show messages and progress bars? \code{\link{readDWD}} will
keep progbar=TRUE for binary files, even if length(file)==1.
DEFAULT: TRUE}

\item{selection}{Optionally read only a subset of the ~24*31=744 files.
Called as \code{f[selection]}. DEFAULT: NULL (ignored)}

\item{\dots}{Further arguments passed to \code{\link{readBin}}}
}
\value{
vector
}
\description{
read gridded radolan binary data.
Intended to be called via \code{\link{readDWD}}.\cr
This does not work correctly yet for the tested data 
(grids_germany/daily/radolan/historical)! Hints are welcome!
}
\examples{
\dontrun{ # Excluded from CRAN checks, but run in localtests

# a single file as example: ----
radfile <- "/daily/radolan/historical/2017/SF201712.tar.gz"
# 204 MB, takes a minute to download:
localfile <- dataDWD(file=radfile, base=gridbase, joinbf=TRUE,
                     dir=localtestdir(), read=FALSE)
rad <- readDWD(localfile, selection=1:10) # no need to read all 24*31=744 files
if(length(rad)!=10) stop("length(rad) should be 10, but is ", length(rad))

# ToDo: make sense of the values, read them correctly!
warning("readDWD.binary does not yet read the binary files correctly.")
raster::plot(raster::raster(matrix(rad[[1]], ncol=900, byrow=TRUE)))


# list of all Files: ----
# radolanfiles <- indexFTP(folder="/daily/radolan", base=gridbase, dir=localtestdir())
radolanfiles <- readLines(localtestdir(file="INDEX_of_DWD_daily_radolan.txt"))
# Files without Bestand folders:
radfiles <- radolanfiles[grepl("historical/..../SF", radolanfiles)]
}
}
\seealso{
\code{\link{readDWD}}\cr
  \url{https://wradlib.org} for much more extensive radar analysis in Python\cr
  Kompositformatbeschreibung at \url{https://www.dwd.de/DE/leistungen/radolan/radolan.html}
  for format description\cr
  \url{https://stats.idre.ucla.edu/r/faq/how-can-i-read-binary-data-into-r} 
  for help I used developing readDWD.binary
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2018
}
