% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDWD.R
\name{readDWD.raster}
\alias{readDWD.raster}
\title{read dwd gridded raster data}
\usage{
readDWD.raster(file, dividebyten, ...)
}
\arguments{
\item{file}{Name of file on harddrive, like e.g. 
DWDdata/grids_germany/seasonal/air_temperature_mean/
16_DJF_grids_germany_seasonal_air_temp_mean_188216.asc.gz}

\item{dividebyten}{Logical: Divide the numerical values by 10?
DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{raster::\link[raster]{raster}}}
}
\value{
\code{raster::\link[raster]{raster}} object
}
\description{
read dwd gridded raster data
}
\seealso{
\code{\link{readDWD}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Dec 2018
}
