% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRadarFile.R
\name{readRadarFile}
\alias{readRadarFile}
\title{read binary radolan radar file}
\usage{
readRadarFile(binfile)
}
\arguments{
\item{binfile}{Name of a single binary file}
}
\value{
A list with dat (matrix/raster???) and meta (list with elements from header)
}
\description{
Read a single binary DWD Radolan file.
             To be used in \code{\link{readDWD.binary}}
}
\examples{
# currently in progress
}
\seealso{
\code{\link{readDWD.binary}}
}
\author{
Original codebase by Henning Rust & Christoph Ritschel at FU Berlin.\cr
        Maintained by Berry Boessenkool, \email{berry-b@gmx.de}, May 2019
}
\keyword{binary}
\keyword{file}
