% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdwd-package.R
\name{rowDisplay}
\alias{rowDisplay}
\title{Create leaflet map popup from data.frame rows}
\usage{
rowDisplay(x)
}
\arguments{
\item{x}{data.frame with colnames}
}
\value{
Vector of character strings, one for each row in x.
}
\description{
Create display character string for leaflet map popup from data.frame rows.
This function is not exported, as it is only internally useful.
A generic version is available in \code{berryFunctions::\link[berryFunctions]{popleaf}}.
}
\seealso{
\code{\link{geoIndex}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Feb 2017
}
\keyword{character}
