% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLocalTests.R
\name{runLocalTests}
\alias{runLocalTests}
\title{run local tests of rdwd}
\usage{
runLocalTests(
  dir_data = locdir(),
  dir_exmpl = berryFunctions::packagePath(file = "misc/ExampleTests"),
  start = 1,
  quiet = rdwdquiet()
)
}
\arguments{
\item{dir_data}{Reusable data location. Preferably not under version control.
DEFAULT: \code{\link[=locdir]{locdir()}}}

\item{dir_exmpl}{Reusable example location. DEFAULT: local directory}

\item{start}{Number to start tests at, helpful for partially successful runs.
DEFAULT: 1}

\item{quiet}{Suppress progress messages? DEFAULT: FALSE through \code{\link[=rdwdquiet]{rdwdquiet()}}}
}
\value{
Time taken to run tests in minutes
}
\description{
Run \code{rdwd} tests on local machine. Due to time-intensive
data downloads, these tests are not run automatically on CRAN.
}
\seealso{
\code{\link[=locdir]{locdir()}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Apr-Oct 2019
}
\keyword{debugging}
