% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDWD.R
\name{selectDWD}
\alias{selectDWD}
\title{Select data from the DWD CDC FTP Server}
\usage{
selectDWD(id = "", name = "", meta = FALSE, index = rdwd:::indexlist,
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  res = "", var = "", time = "")
}
\arguments{
\item{id, name}{Character strings. If both are "", \code{meta} is set to TRUE.
If \code{id} is given, \code{name} is ignored. DEFAULT: ""}

\item{meta}{Logical: return metadata txt file name instead of climate data zip file?
DEFAULT: FALSE}

\item{index}{Index used to select filename, as returned by
\code{\link{index2df}}. To use a current / custom index, use
\code{myIndex <- index2df(indexDWD("/daily/solar"))}
(with desired path, of course). DEFAULT: rdwd:::indexlist}

\item{base}{Main directory of DWD ftp server, defaulting to observed climatic records.
Must be the same \code{base} used to create \code{index}.
DEFAULT: \url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}}

\item{res}{Temporal resolution available at \code{base}, usually one of
\code{c("hourly","daily","monthly")}. DEFAULT: ""}

\item{var}{Weather variable of interest, usually one of
\code{c("kl", "more_precip", "soil_temperature", "solar",
"air_temperature", "cloudiness", "precipitation", "pressure", "sun", "wind")}.
See more in \code{\link{indexlist}}. DEFAULT: ""}

\item{time}{Desired time range. One of
"recent" (data from the last year, up to date usually within a few days) or
"historical" (long time series). DEFAULT: ""}
}
\value{
Character string with file path and name(s) in the format
        "base/res/var/time/filename.zip"
}
\description{
Select files for downloading with \code{\link{dataDWD}}.
Each argument can be a vector, then the rules below are applied element-wise.\cr # ToDo: not yet
If station \code{name} is given, but \code{id} is empty (""),
\code{id} is inferred via \code{findex}.
# ToDo: fileIndex (instead of indexlist), metaIndex. argument names findex, mindex
If \code{res/var/time} are given and valid (existing in \code{findex}),
they are pasted together to form a \code{path}.
Here is an overview of the behaviour in each case of availability:
\tabular{llll}{
case \tab | id \tab | path \tab | action \cr
 1 \tab |  ""  \tab |  ""  \tab | ToDo decide, is now the DEFAULT \cr
 2 \tab | "xx" \tab |  ""  \tab | All file names (across paths) for station "id" \cr
 3 \tab |  ""  \tab | "xx" \tab | meta=TRUE -> name of station description file at path \cr
 4 \tab | "xx" \tab | "xx" \tab | regular single data file name \cr
}
}
\examples{
\dontrun{
selectDWD(id="00386", res="daily", var="kl", time="historical")
selectDWD(id="00386")
selectDWD(id="00386", meta=TRUE)
selectDWD(res="daily", var="kl", time="recent")
selectDWD(var="dummy") # should give an informative error
selectDWD(time="dummy") # this one, too
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{dataDWD}}
}
\keyword{file}

