% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDWD.R
\name{selectDWD}
\alias{selectDWD}
\title{Select data from the DWD CDC FTP Server}
\usage{
selectDWD(name = "", id = "",
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  res = "", var = "", time = "", index = rdwd:::indexlist,
  meta = FALSE, files = FALSE, ziponly = TRUE, ...)
}
\arguments{
\item{name}{Char: station name that will be matched in \code{index} to obtain
\code{id}.  Not yet implemented!  DEFAULT: ""}

\item{id}{Char: station ID with leading zeros, e.g. "00614".
If \code{id} is given, \code{name} is ignored. DEFAULT: ""}

\item{base}{Single char: main directory of DWD ftp server, defaulting to
observed climatic records.
Must be the same \code{base} used to create \code{index}.
DEFAULT: \url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate/}}

\item{res}{Char: temporal resolution available at \code{base}, usually one of
\code{c("hourly","daily","monthly")}. DEFAULT: ""}

\item{var}{Char: Weather variable of interest, usually one of
\code{c("kl", "more_precip", "soil_temperature", "solar",
"air_temperature", "cloudiness", "precipitation", "pressure", "sun", "wind")}.
See more in \code{View(rdwd:::\link{indexlist})}. DEFAULT: ""}

\item{time}{Char: desired time range. One of
"recent" (data from the last year, up to date usually within a few days) or
"historical" (long time series). DEFAULT: ""
ToDo: make time abbreviatable}

\item{index}{Single object: Index used to select filename, as returned by
\code{\link{index2df}}.To use a current / custom index, use
\code{myIndex <- index2df(indexDWD("/daily/solar"))}
(with desired path, of course). DEFAULT: rdwd:::indexlist}

\item{meta}{Logical: return metadata txt file name instead of climate data zip file?
Relevant only in case 4 (path and id given). DEFAULT: FALSE}

\item{files}{Logical: for case 3/4 with given \code{path}: instead of station
metadata, return a list of the currently available files?
This will call code{\link{indexDWD}} with the path, thus
requiring the \code{RCurl} package. DEFAULT: FALSE}

\item{ziponly}{Logical: If \code{files=TRUE}, only return the zip files, not the
description (Beschreibung.txt/.pdf) files? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{indexDWD}} if \code{files=TRUE}}
}
\value{
Character string with file path and name(s) in the format
        "base/res/var/time/filename.zip"
}
\description{
Select files for downloading with \code{\link{dataDWD}}.
All arguments (except for \code{index} and \code{base})
can be a vecor and will be recycled to the maximum length of all arguments.
If that length > 1, the output is a list of filenames.\cr
If station \code{name} is given, but \code{id} is empty (""),
\code{id} is inferred via \code{findex}.
# ToDo: fileIndex (instead of indexlist), metaIndex. argument names findex, mindex
If \code{res/var/time} are given and valid (existing in \code{findex}),
they are pasted together to form a \code{path}.
Here is an overview of the behaviour in each case of availability:
\tabular{llll}{
case \tab | id \tab | path \tab | action \cr
 1 \tab |  ""  \tab |  ""  \tab | ToDo decide, is now the DEFAULT \cr
 2 \tab | "xx" \tab |  ""  \tab | All file names (across paths) for station "id" \cr
 3 \tab |  ""  \tab | "xx" \tab | meta=TRUE -> name of station description file at path \cr
 4 \tab | "xx" \tab | "xx" \tab | regular single data file name \cr
}
}
\examples{
\dontrun{
selectDWD(id="00386", res="daily", var="kl", time="historical")
selectDWD(id="00386", res="daily", var="kl", time="historical", meta=TRUE)
selectDWD(id="00386")
selectDWD(id="00386", meta=TRUE)
selectDWD(res="daily", var="kl", time="recent") # id missing, thus meta<-TRUE
selectDWD(var="dummy") # should give an informative error
selectDWD(time="dummy") # this one, too
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{dataDWD}}
}
\keyword{file}

