% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDWD.R
\name{selectDWD}
\alias{selectDWD}
\title{Select data from the DWD CDC FTP Server}
\usage{
selectDWD(name = "", id = "",
  base = "ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate",
  res = "", var = "", time = "", mindex = rdwd:::metaIndex,
  findex = rdwd:::fileIndex, meta = FALSE, files = FALSE,
  ziponly = TRUE, ...)
}
\arguments{
\item{name}{Char: station name that will be matched in \code{mindex} to obtain
\code{id}. DEFAULT: ""}

\item{id}{Char/Number: station ID with or without leading zeros, e.g. "00614" or 614.
Is internally converted to an integer, because some DWD meta data
Files also contain no leading zeros.
If \code{id} is given, \code{name} is ignored. DEFAULT: ""}

\item{base}{Single char: main directory of DWD ftp server, defaulting to
observed climatic records.
Must be the same \code{base} used to create \code{findex}.
DEFAULT: \url{ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/climate}}

\item{res}{Char: temporal resolution available at \code{base}, usually one of
\code{c("hourly","daily","monthly")}. DEFAULT: ""}

\item{var}{Char: Weather variable of interest, usually one of
\code{c("kl", "more_precip", "soil_temperature", "solar",
"air_temperature", "cloudiness", "precipitation", "pressure", "sun", "wind")}.
See more in \code{View(rdwd:::\link{fileIndex})}. DEFAULT: ""}

\item{time}{Char: desired time range. One of
"recent" (data from the last year, up to date usually within a few days) or
"historical" (long time series). Can be abbreviated (if the first
letter is "r" or "h", full names are used. DEFAULT: ""}

\item{mindex}{Single object: Index used to select \code{id} if \code{name}
is given. DEFAULT: \code{rdwd:::\link{metaIndex}}}

\item{findex}{Single object: Index used to select filename, as returned by
\code{\link{index2df}}.To use a current / custom index, use
\code{myIndex <- index2df(indexDWD("/daily/solar"))}
(with desired path, of course). DEFAULT: \code{rdwd:::\link{fileIndex}}}

\item{meta}{Logical: return metadata txt file name instead of climate data zip file?
Relevant only in case 4 (path and id given). DEFAULT: FALSE}

\item{files}{Logical: for case 3/4 with given \code{path}: instead of station
metadata, return a list of the currently available files?
This will call code{\link{indexDWD}} with the path, thus
requiring the \code{RCurl} package. DEFAULT: FALSE}

\item{ziponly}{Logical: If \code{files=TRUE}, only return the zip files, not the
description (Beschreibung.txt/.pdf) files? DEFAULT: TRUE}

\item{\dots}{Further arguments passed to \code{\link{indexDWD}} if \code{files=TRUE}}
}
\value{
Character string with file path and name(s) in the format
        "base/res/var/time/filename.zip"
}
\description{
Select files for downloading with \code{\link{dataDWD}}.
All arguments (except for \code{mindex}, \code{findex} and \code{base})
can be a vecor and will be recycled to the maximum length of all arguments.
If that length > 1, the output is a list of filenames.\cr
If station \code{name} is given, but \code{id} is empty (""),
\code{id} is inferred via \code{mindex}.
If \code{res/var/time} are given and valid (existing in \code{findex}),
they are pasted together to form a \code{path}.
Here is an overview of the behaviour in each case of availability:
\tabular{llll}{
case \tab | id \tab | path \tab | action \cr
 1 \tab |  ""  \tab |  ""  \tab | ToDo decide, is now the DEFAULT \cr
 2 \tab | "xx" \tab |  ""  \tab | All file names (across paths) for station "id" \cr
 3 \tab |  ""  \tab | "xx" \tab | meta=TRUE -> name of station description file at path \cr
 4 \tab | "xx" \tab | "xx" \tab | regular single data file name \cr
}
}
\examples{
\dontrun{
selectDWD(id="00386", res="daily", var="kl", time="historical")
selectDWD(id="00386", res="daily", var="kl", time="h") # time abbreviatable
selectDWD(id="00386", res="daily", var="kl", time="historical", meta=TRUE)

selectDWD(id="01050", res="daily", var="kl", time=c("r","h")) #vectorizable
selectDWD(id="01050", res=c("daily","monthly"), var="kl") # must give time
selectDWD(id="01050", res=c("daily","monthly"), var="kl", time="r")
# vectorization gives not the outer product, but elementwise comparison:
selectDWD(id="01050", res=c("daily","monthly"), var="kl", time=c("r","h"))

selectDWD(id="01050") # all files in all paths matching id
selectDWD(id="00386", meta=TRUE)
selectDWD(res="daily", var="kl", time="recent") # id missing, thus meta<-TRUE

# See if errors come as expected:
selectDWD(res="dummy", var="dummy") # should give an informative error
selectDWD(res="dummy") # this one, too
selectDWD(res="daily", var="kl") # must give time
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2016
}
\seealso{
\code{\link{dataDWD}}
}
\keyword{file}

