% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateIndexes.R
\name{updateIndexes}
\alias{updateIndexes}
\title{update rdwd indexes}
\usage{
updateIndexes(dwdlocal = FALSE, grdlocal = FALSE, metaforce = NA)
}
\arguments{
\item{dwdlocal}{Read "DWDdata/INDEX_of_DWD_.txt" instead of calling
\code{\link{indexFTP}}? DEFAULT: FALSE}

\item{grdlocal}{Read "DWDdata/INDEX_of_DWD_grids.txt" instead of calling
\code{\link{indexFTP}}? DEFAULT: FALSE}

\item{metaforce}{\code{\link{dataDWD} force} argument for BESCHREIBUNG files. 
DEFAULT: NA (re-download if older than 24 hours)}
}
\value{
\code{\link{checkIndex}} results
}
\description{
This is meant to be called with getwd at the 
 rdwd package source code directory to update the indexes with one single call.
}
\examples{
# number of files at dwdbase
#  25'757 (2017-03-14) 
# 218'593 (2018-03-25)
# 228'830 (2018-11-26)
# 240'737 (2019-02-19)
# 242'584 (2019-03-11)
# 266'860 (2019-05-15)
# 254'446 (2019-05-30)
# 255'252 (2019-07-31)
# 254'925 (2019-09-17)
# 254'943 (2019-10-26)
# 266'841 (2020-03-16)
# 265'712 (2020-04-10)
# 265'712 (2020-04-24)
# 266'106 (2020-06-01)
 
# gridbase
#  49'247 (2019-05-26)
#  49'402 (2019-05-30)
#  54'314 (2019-07-31)
#  56'759 (2019-09-17)
#  58'656 (2019-10-26)
#  30'320 (2020-03-16)
#  31'787 (2020-04-10)
#  32'478 (2020-04-24)
#  34'203 (2020-06-01)

}
\seealso{
\code{\link{createIndex}}, \code{graphics::\link[graphics]{plot}}
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Oct 2019
}
\keyword{data}
\keyword{file}
\keyword{internal}
